requirements.txt
pandas~=2.2.2
numpy~=1.26.4
requests~=2.32.5
urllib3~=2.6.3
python-dotenv~=1.2.1